server/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
black==23.7.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.4
django-admin-interface==0.26.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-colorfield==0.9.0
django-cors-headers==4.0.0
django-countries==7.5.1
django-elasticsearch-dsl==7.3
django-enum-choices==2.1.4
django-environ==0.10.0
django-filter==23.2
django-jazzmin==2.6.0
django-nested-inline==0.4.6
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
mypy==1.5.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.18
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
